GH-2415 allow parsing RDF* using standard sparql json result parser
- support Stardog syntax variant
- support Jena syntax variant
