+++ core/queryresultio/sparqljson/src/main/java/org/eclipse/rdf4j/query/resultio/sparqljson/SPARQLStarResultsJSONConstants.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/org/documents/edl-v10.php.
  *******************************************************************************/
-package org.eclipse.rdf4j.query.resultio.sparqlstarjson;
+package org.eclipse.rdf4j.query.resultio.sparqljson;
 
 import org.eclipse.rdf4j.query.resultio.TupleQueryResultFormat;
