+++ core/queryresultio/sparqljson/src/main/java/org/eclipse/rdf4j/query/resultio/sparqljson/SPARQLStarResultsJSONWriter.java
@@ -1,0 +1,65 @@
+/*******************************************************************************
+ * Copyright (c) 2020 Eclipse RDF4J contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Distribution License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ *******************************************************************************/
+package org.eclipse.rdf4j.query.resultio.sparqljson;
+
+import java.io.IOException;
+import java.io.OutputStream;
+
+import org.eclipse.rdf4j.model.Triple;
+import org.eclipse.rdf4j.model.Value;
+import org.eclipse.rdf4j.query.QueryResultHandlerException;
+import org.eclipse.rdf4j.query.resultio.TupleQueryResultFormat;
+import org.eclipse.rdf4j.query.resultio.TupleQueryResultWriter;
+
+/**
+ * Writer for SPARQL* JSON results. This is equivalent to the SPARQL JSON writer with the addition of support for RDF*
+ * triples. See {@link SPARQLStarResultsJSONConstants} for a description of the RDF* extension.
+ *
+ * @author Pavel Mihaylov
+ */
+public class SPARQLStarResultsJSONWriter extends SPARQLResultsJSONWriter implements TupleQueryResultWriter {
+	public SPARQLStarResultsJSONWriter(OutputStream out) {
+		super(out);
+	}
+
+	@Override
+	public TupleQueryResultFormat getTupleQueryResultFormat() {
+		return SPARQLStarResultsJSONConstants.QUERY_RESULT_FORMAT;
+	}
+
+	@Override
+	public TupleQueryResultFormat getQueryResultFormat() {
+		return getTupleQueryResultFormat();
+	}
+
+	@Override
+	protected void writeValue(Value value) throws IOException, QueryResultHandlerException {
+		if (value instanceof Triple) {
+			jg.writeStartObject();
+
+			jg.writeStringField(AbstractSPARQLJSONParser.TYPE, SPARQLStarResultsJSONConstants.TRIPLE);
+
+			jg.writeObjectFieldStart(AbstractSPARQLJSONParser.VALUE);
+
+			jg.writeFieldName(SPARQLStarResultsJSONConstants.SUBJECT);
+			writeValue(((Triple) value).getSubject());
+
+			jg.writeFieldName(SPARQLStarResultsJSONConstants.PREDICATE);
+			writeValue(((Triple) value).getPredicate());
+
+			jg.writeFieldName(SPARQLStarResultsJSONConstants.OBJECT);
+			writeValue(((Triple) value).getObject());
+
+			jg.writeEndObject();
+
+			jg.writeEndObject();
+		} else {
+			super.writeValue(value);
+		}
+	}
+}
