+++ core/queryresultio/sparqljson/src/main/resources/META-INF/services/org.eclipse.rdf4j.query.resultio.TupleQueryResultParserFactory
@@ -1,2 +1,2 @@
 org.eclipse.rdf4j.query.resultio.sparqljson.SPARQLResultsJSONParserFactory
-org.eclipse.rdf4j.query.resultio.sparqlstarjson.SPARQLStarResultsJSONParserFactory
+org.eclipse.rdf4j.query.resultio.sparqljson.SPARQLStarResultsJSONParserFactory
