+++ core/queryresultio/sparqljson/src/main/resources/META-INF/services/org.eclipse.rdf4j.query.resultio.TupleQueryResultWriterFactory
@@ -1,2 +1,2 @@
 org.eclipse.rdf4j.query.resultio.sparqljson.SPARQLResultsJSONWriterFactory
-org.eclipse.rdf4j.query.resultio.sparqlstarjson.SPARQLStarResultsJSONWriterFactory
+org.eclipse.rdf4j.query.resultio.sparqljson.SPARQLStarResultsJSONWriterFactory
