+++ core/queryresultio/sparqljson/src/test/java/org/eclipse/rdf4j/query/resultio/sparqljson/SPARQLStarJSONTupleBackgroundTest.java
@@ -5,16 +5,24 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/org/documents/edl-v10.php.
  *******************************************************************************/
-package org.eclipse.rdf4j.query.resultio.sparqlstarjson;
+package org.eclipse.rdf4j.query.resultio.sparqljson;
 
+import java.io.IOException;
+import java.io.InputStream;
+
+import org.eclipse.rdf4j.query.TupleQueryResult;
+import org.eclipse.rdf4j.query.TupleQueryResultHandlerException;
 import org.eclipse.rdf4j.query.resultio.AbstractQueryResultIOTupleTest;
 import org.eclipse.rdf4j.query.resultio.BooleanQueryResultFormat;
+import org.eclipse.rdf4j.query.resultio.QueryResultIO;
+import org.eclipse.rdf4j.query.resultio.QueryResultParseException;
 import org.eclipse.rdf4j.query.resultio.TupleQueryResultFormat;
+import org.eclipse.rdf4j.query.resultio.UnsupportedQueryResultFormatException;
 
 /**
  * @author Pavel Mihaylov
  */
-public class SPARQLStarJSONTupleTest extends AbstractQueryResultIOTupleTest {
+public class SPARQLStarJSONTupleBackgroundTest extends AbstractQueryResultIOTupleTest {
 	@Override
 	protected String getFileName() {
 		return "test.srjs";
@@ -28,5 +36,11 @@ public class SPARQLStarJSONTupleTest extends AbstractQueryResultIOTupleTest {
 	@Override
 	protected BooleanQueryResultFormat getMatchingBooleanFormatOrNull() {
 		return BooleanQueryResultFormat.JSON;
+	}
+
+	@Override
+	protected TupleQueryResult parseTupleInternal(TupleQueryResultFormat format, InputStream in) throws IOException,
+			QueryResultParseException, TupleQueryResultHandlerException, UnsupportedQueryResultFormatException {
+		return QueryResultIO.parseTupleBackground(in, format);
 	}
 }
